@@ -8,13 +8,13 @@ import org.testcontainers.containers.DockerComposeContainer;
 import org.testcontainers.containers.MongoDBContainer;
 
 import java.util.ArrayList;
+import java.util.stream.Stream;
 
 import static com.webflux.api.core.config.utils.BlockhoundUtils.blockhoundInstallWithSpecificAllowedCalls;
 import static com.webflux.api.core.config.utils.RestAssureSpecs.requestSpecs;
 import static com.webflux.api.core.config.utils.RestAssureSpecs.responseSpecs;
-import static java.lang.Math.min;
+import static java.lang.String.valueOf;
 import static java.lang.System.out;
-import static java.util.stream.Stream.of;
 
 @Slf4j
 public class TestUtils {
@@ -25,9 +25,8 @@ public class TestUtils {
     requestSpecs();
     responseSpecs();
     blockhoundInstallWithSpecificAllowedCalls();
-    //        blockhoundInstallWithAllAllowedCalls();
+    //  blockhoundInstallWithAllAllowedCalls();
   }
-
 
   @AfterAll
   public static void globalAfterAll() {
@@ -35,372 +34,410 @@ public class TestUtils {
     RestAssuredWebTestClient.reset();
   }
 
+  public static void globalTestMessage(String content, String testType) {
 
-  public static void globalTestMessage(String subTitle, String testType) {
+    final String error =
+         "Error: Provide TestInfo testInfo.getTestMethod().toString()";
+    // @formatter:off
+    var contentEdited =
+         Stream.of(content)
+               .map(check -> check.contains("repetition") ? error : check)
+               .map(text -> {
+                 var check = text.contains("()]") & ! text.contains("repetition");
+                 if (check) return text
+                      .transform(txt -> txt.replace("()]", ""))
+                      .transform(txt -> txt.substring(text.lastIndexOf(".") + 1))
+                      .transform(txt -> txt.substring(0, 1).toUpperCase() + txt.substring(1));
+                 return text;
+               })
+               .findAny()
+               .get();
+    // @formatter:on
 
+    String title = switch (testType.toLowerCase()) {
+      case "class-start" -> " STARTING TEST-CLASS...";
+      case "class-end" -> "...FINISHED TEST-CLASS ";
+      case "method-start" -> "STARTING TEST-METHOD...";
+      case "method-end" -> "...FINISHED TEST-METHOD";
+      default -> "";
+    };
 
-    if (subTitle.contains("repetition"))
-      subTitle = "Error: Provide TestInfo testInfo.getTestMethod().toString()";
-
-    if (subTitle.contains("()]")) {
-      subTitle = subTitle.replace("()]", "");
-      subTitle = subTitle.substring(subTitle.lastIndexOf(".") + 1);
-      subTitle = subTitle.substring(0, 1)
-                         .toUpperCase() + subTitle.substring(1);
-    }
-
-    String title;
-
-    switch (testType.toLowerCase()) {
-      case "class-start":
-        title = " STARTING TEST-CLASS...";
-        break;
-      case "class-end":
-        title = "...FINISHED TEST-CLASS ";
-        break;
-      case "method-start":
-        title = "STARTING TEST-METHOD...";
-        break;
-      case "method-end":
-        title = "...FINISHED TEST-METHOD";
-        break;
-      default:
-        title = "";
-    }
-
-
-    out.printf(
-
-         "╔════════════════════════════════════════════════════════════════════╗\n" + "║         "
-         + "              %s                                           ║\n" + "║ --> Name: %s " + "%38s%n" + "╚════════════════════════════════════════════════════════════════════╝\n",
-         title, subTitle, "║"
-    );
+    ConsolePanel.simplePanel(title, contentEdited);
   }
 
+  public static void globalContainerMessage(
+       MongoDBContainer container,
+       String typeTestMessage) {
 
-  public static void globalContainerMessage(MongoDBContainer container, String typeTestMessage) {
+    // @formatter:off
+    if (container == null) return;
 
-    if (container != null) {
-      String title;
-      switch (typeTestMessage.toLowerCase()) {
-        case "container-start":
-          title = "STARTING TEST-CONTAINER...";
-          break;
-        case "container-end":
-          title = "...FINISHED TEST-CONTAINER";
-          break;
-        case "container-state":
-          title = "  ...TEST'S TC-CONTAINER  ";
-          break;
-        default:
-          title = "";
-      }
+    String title = switch (typeTestMessage.toLowerCase()) {
+      case "container-start" -> "STARTING TEST-CONTAINER...";
+      case "container-end" -> "...FINISHED TEST-CONTAINER";
+      case "container-state" -> "  ...TEST'S TC-CONTAINER  ";
+      default -> "";
+    };
 
+    ConsolePanel.simplePanel(
+         title,
+         container.getContainerName().transform("Name: "::concat),
+         container.getReplicaSetUrl().transform("Url: "::concat),
+         "Running: ".concat(valueOf(container.isRunning())));
+    // @formatter:on
 
-      out.printf(
-           "╔═══════════════════════════════════════════════════════════════════════╗\n" + "║ " + "-->" + " Name: %s\n" + "║ --> Url: %s\n" + "║ --> Running: %s\n" + "╚═══════════════════════════════════════════════════════════════════════╝\n\n",
-           title, container.getContainerName(), container.getReplicaSetUrl(), container.isRunning()
-      );
-    }
   }
 
+  public static void globalComposeServiceContainerMessage(
+       DockerComposeContainer<?> compose,
+       String service,
+       Integer port) {
 
-  public static void globalComposeServiceContainerMessage(DockerComposeContainer<?> compose,
-                                                          String service, Integer port) {
+    // @formatter:off
+    if (compose == null) return;
 
-    if (compose != null) {
-      out.printf(
-
-           "╔═══════════════════════════════════════════════════════════════════════\n" + "║     "
-           + "                      %s                        ║\n" + "║ --> Service: %s\n" + "║ " + "-->" + " Host: %s\n" + "║ --> Port: %s\n" + "║ --> Created: %s\n" + "║ --> Running: " + "%s\n" + "╚═══════════════════════════════════════════════════════════════════════\n\n",
-           "TC-CONTAINER-COMPOSE", service, compose.getServiceHost(service, port),
-           compose.getServicePort(service, port), compose.getContainerByServiceName(service + "_1")
-                                                         .get()
-                                                         .isCreated(),
-           compose.getContainerByServiceName(service + "_1")
-                  .get()
-                  .isRunning()
-      );
-    }
+    var container = compose.getContainerByServiceName(service + "_1").get();
+    ConsolePanel
+         .simplePanel("TC-CONTAINER-COMPOSE",
+                      service,
+                      compose.getServiceHost(service, port),
+                      compose.getServicePort(service, port).toString(),
+                      "Created: ".concat(valueOf(container.isCreated())),
+                      "Running: ".concat(valueOf(container.isRunning()))
+         );
+    // @formatter:on
   }
 
   private static class ConsolePanel {
 
-    public static void simplePanel(String... texts) {
-      // @formatter:off
-      panel(45,
-            5,
-            1,
-            1,
-            Border.DOUBLE,
-            Border.DOUBLE,
-            Border.THIN,
-            Border.THIN,
-            true,
-            true, texts);
-      // @formatter:on
-    }
-
-    public static void simplePanelScalable(int scale, String... texts) {
-      // @formatter:off
-      panel(scale,
-            5,
-            1,
-            1,
-            Border.DOUBLE,
-            Border.DOUBLE,
-            Border.THIN,
-            Border.THIN,
-            true,
-            true,
-            texts
-      );
-      // @formatter:on
-    }
-
     // @formatter:off
-    public static void panel(
-         int scale,
-         int margin,
-         int upSpace,
-         int downSpace,
-         Border cornersFormat, Border centerMarksFormat,
-         Border horizontalLinesFormat,
-         Border verticalLinesFormat,
-         boolean uppercaseTitle,
-         boolean centralizeTitle,
-         String... titleAndOthers) {
-      // @formatter:on
+  public static void simplePanel(String... texts) {
+    panel(45,
+          5,
+          1,
+          1,
+          Border.DOUBLE,
+          Border.DOUBLE,
+          Border.THIN,
+          Border.THIN,
+          true,
+          true, texts);
+  }
 
-      var estimatedAdjustmentFactor = 3;
-      var title = titleAndOthers[0];
-      var marginTitle = scale - (title.length() / 2) - estimatedAdjustmentFactor;
-      var formattedTexts = of(titleAndOthers).map(
-                                                  item -> item.equals(title) && centralizeTitle ?
-                                                       " ".repeat(marginTitle) + title : item)
-                                             .map(item -> item.equals(
-                                                  title) && uppercaseTitle ? item.toUpperCase() :
-                                                  item)
-                                             .toArray();
+  public static void simplePanelScalable(int scale, String... texts) {
+    panel(scale,
+          5,
+          1,
+          1,
+          Border.DOUBLE,
+          Border.DOUBLE,
+          Border.THIN,
+          Border.THIN,
+          true,
+          true,
+          texts);
+  }
 
-      var marginLimitedBySize = min(margin, scale);
+  public static void panel(
+       int scale,
+       int margin,
+       int upSpace,
+       int downSpace,
+       Border cornersFormat, Border centerMarksFormat,
+       Border horizontalLinesFormat,
+       Border verticalLinesFormat,
+       boolean uppercaseTitle,
+       boolean centralizeTitle,
+       String... titleAndOthers)
+  {
 
-      // scale + margin discrepacies eliminated
-      if (marginLimitedBySize % 2 != 0) -- marginLimitedBySize;
-      if (scale % 2 != 0) ++ scale;
+    var estimatedAdjustmentFactor = 3;
+    var title = titleAndOthers[0];
+    var marginTitle = scale - (title.length() / 2) - estimatedAdjustmentFactor;
+    var formattedTexts =
+         Stream.of(titleAndOthers)
+               .map(
+                    item -> item.equals(title) && centralizeTitle ?
+                         " ".repeat(marginTitle) + title :
+                         item
+               )
+               .map(item -> item.equals(
+                    title) && uppercaseTitle ?
+                    item.toUpperCase() :
+                    item
+               )
+               .toArray();
 
-      int fullSize = (scale * 2) - marginLimitedBySize;
-      if (fullSize % 2 == 0) ++ fullSize;
-      else -- fullSize;
+    var marginLimitedBySize = Math.min(margin, scale);
 
-      var whitespaceMargin = " ".repeat(marginLimitedBySize);
-      var externalUpSpaces = "\n".repeat(upSpace);
-      var externalBottomSpaces = "\n".repeat(downSpace);
+    // scale + margin discrepacies eliminated
+    if (marginLimitedBySize % 2 != 0) -- marginLimitedBySize;
+    if (scale % 2 != 0) ++ scale;
 
-      var upperFace = upperLine(scale, cornersFormat, centerMarksFormat, horizontalLinesFormat);
-      var divider = middleLine(scale, cornersFormat, centerMarksFormat, horizontalLinesFormat);
-      var bottomFace = bottomLine(scale, cornersFormat, centerMarksFormat, horizontalLinesFormat);
-      var faceLine = faceLine(verticalLinesFormat);
+    int fullSize = (scale * 2) - marginLimitedBySize;
+    if (fullSize % 2 == 0) ++ fullSize;
+    else -- fullSize;
 
-      var titleTextArea = String.valueOf(fullSize);
-      var textPreparation = new StringBuilder();
-      textPreparation.append(externalUpSpaces)
-                     .append(upperFace)
+    var whitespaceMargin = " ".repeat(marginLimitedBySize);
+    var externalUpSpaces = "\n".repeat(upSpace);
+    var externalBottomSpaces = "\n".repeat(downSpace);
+
+    var upperFace =
+         upperLine(scale, cornersFormat, centerMarksFormat, horizontalLinesFormat);
+    var divider =
+         middleLine(scale, cornersFormat, centerMarksFormat, horizontalLinesFormat);
+    var bottomFace =
+         bottomLine(scale, cornersFormat, centerMarksFormat, horizontalLinesFormat);
+    var faceLine = faceLine(verticalLinesFormat);
+
+    var titleTextArea = String.valueOf(fullSize);
+    var textPreparation = new StringBuilder();
+    textPreparation.append(externalUpSpaces)
+                   .append(upperFace)
+                   .append(faceLine)
+                   .append("%s%%-%ss".formatted(whitespaceMargin, titleTextArea))
+                   .append(faceLine)
+                   .append("\n")
+                   .append(divider);
+
+    // "-1" Because the first element in the Array was used as title
+    // The discont-number in bodyTextArea/fullsize, subtract the size of "ordinal-ASC" and ") "
+    var bodyTextArea = String.valueOf(fullSize - 4);
+    var topicEnumeration = 0;
+    var ordinalSymbolEnumerator = '\u2070';
+    for (int i = formattedTexts.length - 1; i > 0; i--) {
+      ++ topicEnumeration;
+      textPreparation.append(faceLine)
+                     .append(
+                          "%s%s%s) %%-%ss"
+                               .formatted(
+                                    whitespaceMargin,
+                                    topicEnumeration,
+                                    ordinalSymbolEnumerator,
+                                    bodyTextArea))
                      .append(faceLine)
-                     .append("%s%%-%ss".formatted(whitespaceMargin, titleTextArea))
-                     .append(faceLine)
-                     .append("\n")
-                     .append(divider);
+                     .append("\n");
+    }
+    textPreparation.append(bottomFace)
+                   .append(externalBottomSpaces);
 
-      // "-1" Because the first element in the Array was used as title
-      // The discont-number in bodyTextArea/fullsize, subtract the size of "ordinal-ASC" and ") "
-      var bodyTextArea = String.valueOf(fullSize - 4);
-      var topicEnumeration = 0;
-      var ordinalSymbolEnumerator = '\u2070';
-      for (int i = formattedTexts.length - 1; i > 0; i--) {
-        ++ topicEnumeration;
-        textPreparation.append(faceLine)
-                       .append("%s%s%s) %%-%ss".formatted(whitespaceMargin, topicEnumeration,
-                                                          ordinalSymbolEnumerator, bodyTextArea
-                       ))
-                       .append(faceLine)
-                       .append("\n");
-      }
-      textPreparation.append(bottomFace)
-                     .append(externalBottomSpaces);
-      out.printf(textPreparation.toString(), formattedTexts);
+    out.printf(textPreparation.toString(), formattedTexts);
+  }
+
+  private static String generateLine(char baseChar, int scale, char BASE_LINE)
+  {
+
+    return String.valueOf(baseChar)
+                 .repeat(scale)
+                 .replace(baseChar, BASE_LINE);
+  }
+
+  private static String upperLine(
+       int scale,
+       Border corner,
+       Border centerMark,
+       Border line)
+  {
+    ArrayList<Character> borderStylingItems = new ArrayList<>();
+    switch (corner) {
+      case BOLD -> {
+        borderStylingItems.add(BoldFont.UPPER_LEFT_CORNER.code);
+        borderStylingItems.add(BoldFont.UPPER_RIGHT_CORNER.code);}
+
+      case THIN -> {
+        borderStylingItems.add(ThinFont.UPPER_LEFT_CORNER.code);
+        borderStylingItems.add(ThinFont.UPPER_RIGHT_CORNER.code);}
+
+      case DOUBLE -> {
+        borderStylingItems.add(DoubleFont.UPPER_LEFT_CORNER.code);
+        borderStylingItems.add(DoubleFont.UPPER_RIGHT_CORNER.code);}
     }
 
-    private static String generateLine(char baseChar, int scale, char BASE_LINE) {
-
-      return String.valueOf(baseChar)
-                   .repeat(scale)
-                   .replace(baseChar, BASE_LINE);
+    switch (centerMark) {
+      case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
+      case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
+      case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
     }
 
-    private static String upperLine(int scale, Border corner, Border centerMark, Border line) {
-
-      ArrayList<Character> borderStylingItems = new ArrayList<>();
-      switch (corner) {
-        case BOLD -> {
-          borderStylingItems.add(BoldFont.UPPER_LEFT_CORNER.code);
-          borderStylingItems.add(BoldFont.UPPER_RIGHT_CORNER.code);
-        }
-        case THIN -> {
-          borderStylingItems.add(ThinFont.UPPER_LEFT_CORNER.code);
-          borderStylingItems.add(ThinFont.UPPER_RIGHT_CORNER.code);
-        }
-        case DOUBLE -> {
-          borderStylingItems.add(DoubleFont.UPPER_LEFT_CORNER.code);
-          borderStylingItems.add(DoubleFont.UPPER_RIGHT_CORNER.code);
-        }
-      }
-
-      switch (centerMark) {
-        case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
-        case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
-        case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
-      }
-
-      switch (line) {
-        case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
-        case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
-        case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
-      }
-
-      var baseline = generateLine('_', scale, borderStylingItems.get(3));
-
-      return borderStylingItems.get(0) + baseline + borderStylingItems.get(
-           2) + baseline + borderStylingItems.get(1) + "\n";
+    switch (line) {
+      case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
+      case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
+      case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
     }
 
-    private static String middleLine(int scale, Border corner, Border centerMark,
-                                     Border baseLine) {
+    var baseline =
+         generateLine('_', scale, borderStylingItems.get(3));
 
-      ArrayList<Character> borderStylingItems = new ArrayList<>();
-      switch (corner) {
-        case BOLD -> {
-          borderStylingItems.add(BoldFont.MIDDLE_LEFT.code);
-          borderStylingItems.add(BoldFont.MIDDLE_RIGHT.code);
-        }
-        case THIN -> {
-          borderStylingItems.add(ThinFont.MIDDLE_LEFT.code);
-          borderStylingItems.add(ThinFont.MIDDLE_RIGHT.code);
-        }
-        case DOUBLE -> {
-          borderStylingItems.add(DoubleFont.MIDDLE_LEFT.code);
-          borderStylingItems.add(DoubleFont.MIDDLE_RIGHT.code);
-        }
-      }
+    return "%s%s%s%s%s\n"
+         .formatted(
+              borderStylingItems.get(0),
+              baseline,
+              borderStylingItems.get(2),
+              baseline,
+              borderStylingItems.get(1));
+  }
 
-      switch (centerMark) {
-        case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
-        case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
-        case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
-      }
+  private static String middleLine(
+       int scale,
+       Border corner,
+       Border centerMark,
+       Border baseLine)
+  {
 
-      switch (baseLine) {
-        case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
-        case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
-        case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
-      }
+    ArrayList<Character> borderStylingItems = new ArrayList<>();
 
-      var divider = generateLine('_', scale, borderStylingItems.get(3));
+    switch (corner) {
+      case BOLD -> {
+        borderStylingItems.add(BoldFont.MIDDLE_LEFT.code);
+        borderStylingItems.add(BoldFont.MIDDLE_RIGHT.code);}
 
-      return borderStylingItems.get(0) + divider + borderStylingItems.get(
-           2) + divider + borderStylingItems.get(1) + "\n";
+      case THIN -> {
+        borderStylingItems.add(ThinFont.MIDDLE_LEFT.code);
+        borderStylingItems.add(ThinFont.MIDDLE_RIGHT.code);}
+
+      case DOUBLE -> {
+        borderStylingItems.add(DoubleFont.MIDDLE_LEFT.code);
+        borderStylingItems.add(DoubleFont.MIDDLE_RIGHT.code);}
     }
 
-    private static String bottomLine(int scale, Border corner, Border centerMark, Border baseLine) {
-
-      ArrayList<Character> borderStylingItems = new ArrayList<>();
-      switch (corner) {
-        case BOLD -> {
-          borderStylingItems.add(BoldFont.LOWER_LEFT_CORNER.code);
-          borderStylingItems.add(BoldFont.LOWER_RIGHT_CORNER.code);
-        }
-        case THIN -> {
-          borderStylingItems.add(ThinFont.LOWER_LEFT_CORNER.code);
-          borderStylingItems.add(ThinFont.LOWER_RIGHT_CORNER.code);
-        }
-        case DOUBLE -> {
-          borderStylingItems.add(DoubleFont.LOWER_LEFT_CORNER.code);
-          borderStylingItems.add(DoubleFont.LOWER_RIGHT_CORNER.code);
-        }
-      }
-
-      switch (centerMark) {
-        case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
-        case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
-        case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
-      }
-
-      switch (baseLine) {
-        case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
-        case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
-        case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
-      }
-
-      var baseline = generateLine('_', scale, borderStylingItems.get(3));
-
-      return borderStylingItems.get(0) + baseline + borderStylingItems.get(
-           2) + baseline + borderStylingItems.get(1) + "\n";
+    switch (centerMark) {
+      case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
+      case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
+      case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
     }
 
-    private static Character faceLine(Border corner) {
-
-      return switch (corner) {
-        case BOLD -> BoldFont.FACE_LINE.code;
-        case THIN -> ThinFont.FACE_LINE.code;
-        case DOUBLE -> DoubleFont.FACE_LINE.code;
-      };
+    switch (baseLine) {
+      case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
+      case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
+      case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
     }
 
-    private enum BoldFont {
+    var divider =
+         generateLine('_', scale, borderStylingItems.get(3));
 
-      FACE_LINE('\u2503'), BASE_LINE('\u2501'), UPPER_LEFT_CORNER('\u250F'), UPPER_RIGHT_CORNER(
-           '\u2513'), MIDDLE_LEFT('\u2523'), MIDDLE_RIGHT('\u252B'), LOWER_LEFT_CORNER(
-           '\u2517'), LOWER_RIGHT_CORNER('\u251B');
+    return "%s%s%s%s%s\n"
+         .formatted(
+              borderStylingItems.get(0),
+              divider,
+              borderStylingItems.get(2),
+              divider,
+              borderStylingItems.get(1));
+  }
 
-      private final char code;
+  private static String bottomLine(
+       int scale,
+       Border corner,
+       Border centerMark,
+       Border baseLine)
+  {
 
-      BoldFont(char code) {
+    ArrayList<Character> borderStylingItems = new ArrayList<>();
+    switch (corner) {
+      case BOLD -> {
+        borderStylingItems.add(BoldFont.LOWER_LEFT_CORNER.code);
+        borderStylingItems.add(BoldFont.LOWER_RIGHT_CORNER.code);}
 
-        this.code = code;
-      }
+      case THIN -> {
+        borderStylingItems.add(ThinFont.LOWER_LEFT_CORNER.code);
+        borderStylingItems.add(ThinFont.LOWER_RIGHT_CORNER.code);}
+
+      case DOUBLE -> {
+        borderStylingItems.add(DoubleFont.LOWER_LEFT_CORNER.code);
+        borderStylingItems.add(DoubleFont.LOWER_RIGHT_CORNER.code);}
     }
 
-    private enum ThinFont {
-
-      FACE_LINE('\u2502'), BASE_LINE('\u2500'), UPPER_LEFT_CORNER('\u250C'), UPPER_RIGHT_CORNER(
-           '\u2510'), MIDDLE_LEFT('\u252C'), MIDDLE_RIGHT('\u2524'), LOWER_LEFT_CORNER(
-           '\u2514'), LOWER_RIGHT_CORNER('\u2518');
-
-      private final char code;
-
-      ThinFont(char code) {
-
-        this.code = code;
-      }
+    switch (centerMark) {
+      case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
+      case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
+      case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
     }
 
-    private enum DoubleFont {
-
-      FACE_LINE('\u2551'), BASE_LINE('\u2550'), UPPER_LEFT_CORNER('\u2554'), UPPER_RIGHT_CORNER(
-           '\u2557'), MIDDLE_LEFT('\u2560'), MIDDLE_RIGHT('\u2563'), LOWER_LEFT_CORNER(
-           '\u255A'), LOWER_RIGHT_CORNER('\u255D');
-
-      private final char code;
-
-      DoubleFont(char code) {
-
-        this.code = code;
-      }
+    switch (baseLine) {
+      case BOLD -> borderStylingItems.add(BoldFont.BASE_LINE.code);
+      case THIN -> borderStylingItems.add(ThinFont.BASE_LINE.code);
+      case DOUBLE -> borderStylingItems.add(DoubleFont.BASE_LINE.code);
     }
 
-    private enum Border {
-      BOLD, THIN, DOUBLE
+    var baseline =
+         generateLine('_', scale, borderStylingItems.get(3));
+
+    return "%s%s%s%s%s\n"
+         .formatted(
+              borderStylingItems.get(0),
+              baseline,
+              borderStylingItems.get(2),
+              baseline,
+              borderStylingItems.get(1));
+  }
+
+  private static Character faceLine(Border corner)
+  {
+
+    return switch (corner) {
+      case BOLD -> BoldFont.FACE_LINE.code;
+      case THIN -> ThinFont.FACE_LINE.code;
+      case DOUBLE -> DoubleFont.FACE_LINE.code;
+    };
+  }
+
+  private enum BoldFont
+  {
+    FACE_LINE('\u2503'),
+    BASE_LINE('\u2501'),
+    UPPER_LEFT_CORNER('\u250F'),
+    UPPER_RIGHT_CORNER('\u2513'),
+    MIDDLE_LEFT('\u2523'),
+    MIDDLE_RIGHT('\u252B'),
+    LOWER_LEFT_CORNER('\u2517'),
+    LOWER_RIGHT_CORNER('\u251B');
+
+    private final char code;
+
+    BoldFont(char code) {
+      this.code = code;
     }
   }
 
+  private enum ThinFont
+  {
+    FACE_LINE('\u2502'),
+    BASE_LINE('\u2500'),
+    UPPER_LEFT_CORNER('\u250C'),
+    UPPER_RIGHT_CORNER('\u2510'),
+    MIDDLE_LEFT('\u252C'),
+    MIDDLE_RIGHT('\u2524'),
+    LOWER_LEFT_CORNER('\u2514'),
+    LOWER_RIGHT_CORNER('\u2518');
+
+    private final char code;
+
+    ThinFont(char code) {
+      this.code = code;
+    }
+  }
+
+  private enum DoubleFont
+  {
+    FACE_LINE('\u2551'),
+    BASE_LINE('\u2550'),
+    UPPER_LEFT_CORNER('\u2554'),
+    UPPER_RIGHT_CORNER('\u2557'),
+    MIDDLE_LEFT('\u2560'),
+    MIDDLE_RIGHT('\u2563'),
+    LOWER_LEFT_CORNER('\u255A'),
+    LOWER_RIGHT_CORNER('\u255D');
+
+    private final char code;
+
+    DoubleFont(char code) {
+      this.code = code;
+    }
+  }
+
+  private enum Border
+  {
+    BOLD, THIN, DOUBLE
+  }
+  // @formatter:on
+  }
 }
